requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.3.3
    # via jupyter-server
argon2-cffi==21.1.0
    # via
    #   jupyter-server
    #   notebook
attrs==21.2.0
    # via jsonschema
babel==2.9.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
bleach==4.1.0
    # via nbconvert
certifi==2021.10.8
    # via requests
cffi==1.14.6
    # via argon2-cffi
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via pip-tools
cycler==0.10.0
    # via matplotlib
debugpy==1.5.0
    # via ipykernel
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via nbconvert
dostoevsky==0.6.0
    # via -r requirements.in
entrypoints==0.3
    # via
    #   jupyter-client
    #   jupyterlab-server
    #   nbconvert
fasttext==0.9.2
    # via dostoevsky
idna==3.3
    # via
    #   anyio
    #   requests
ipykernel==6.4.1
    # via notebook
ipython==7.28.0
    # via
    #   ipykernel
    #   jupyterlab
ipython-genutils==0.2.0
    # via
    #   ipykernel
    #   jupyter-server
    #   nbformat
    #   notebook
jedi==0.18.0
    # via ipython
jinja2==3.0.2
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   notebook
json5==0.9.6
    # via jupyterlab-server
jsonschema==4.1.0
    # via
    #   jupyterlab-server
    #   nbformat
jupyter-client==7.0.6
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
    #   notebook
jupyter-core==4.8.1
    # via
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbconvert
    #   nbformat
    #   notebook
jupyter-server==1.11.1
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
jupyterlab==3.1.18
    # via -r requirements.in
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-server==2.8.2
    # via jupyterlab
kiwisolver==1.3.2
    # via matplotlib
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.3
    # via
    #   seaborn
    #   wordcloud
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclassic==0.3.2
    # via jupyterlab
nbclient==0.5.4
    # via nbconvert
nbconvert==6.2.0
    # via
    #   jupyter-server
    #   notebook
nbformat==5.1.3
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
notebook==6.4.4
    # via nbclassic
numpy==1.21.2
    # via
    #   fasttext
    #   matplotlib
    #   pandas
    #   scipy
    #   seaborn
    #   wordcloud
packaging==21.0
    # via
    #   bleach
    #   jupyterlab
    #   jupyterlab-server
pandas==1.3.3
    # via
    #   -r requirements.in
    #   seaborn
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.2
    # via jedi
pep517==0.11.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.2
    # via
    #   matplotlib
    #   wordcloud
pip-tools==6.4.0
    # via -r requirements.in
prometheus-client==0.11.0
    # via
    #   jupyter-server
    #   notebook
prompt-toolkit==3.0.20
    # via ipython
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pybind11==2.8.0
    # via fasttext
pycparser==2.20
    # via cffi
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-obscene-words-filter==0.1.6
    # via -r requirements.in
pytz==2021.3
    # via
    #   babel
    #   pandas
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   jupyter-server
    #   notebook
razdel==0.5.0
    # via
    #   -r requirements.in
    #   dostoevsky
requests==2.26.0
    # via
    #   jupyterlab-server
    #   requests-unixsocket
requests-unixsocket==0.2.0
    # via jupyter-server
scipy==1.7.1
    # via seaborn
seaborn==0.11.2
    # via -r requirements.in
send2trash==1.8.0
    # via
    #   jupyter-server
    #   notebook
six==1.16.0
    # via
    #   bleach
    #   cycler
    #   python-dateutil
sniffio==1.2.0
    # via anyio
stop-words==2018.7.23
    # via -r requirements.in
terminado==0.12.1
    # via
    #   jupyter-server
    #   notebook
testpath==0.5.0
    # via nbconvert
tomli==1.2.1
    # via pep517
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.62.3
    # via -r requirements.in
traitlets==5.1.0
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
urllib3==1.26.7
    # via
    #   requests
    #   requests-unixsocket
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websocket-client==1.2.1
    # via jupyter-server
wheel==0.37.0
    # via pip-tools
wordcloud==1.8.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
